backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.4.0
mysqlclient==2.2.7
pillow==11.1.0
PyJWT==1.7.1
sqlparse==0.5.3
tzdata==2025.1
